requirements.txt
torch
torchvision
transformers
opencv-python
pillow
numpy
supervision
huggingface_hub
python-dotenv
